nonce is not 0x00 (witnessUtxo isconfidential test func)
